+++ DiligentEngine/Structs/BuildBLASAttribs.cs
@@ -32,7 +32,7 @@ namespace DiligentEngine
         public List<BLASBuildTriangleData> pTriangleData { get; set; }
         public List<BLASBuildBoundingBoxData> pBoxData { get; set; }
         public IBuffer pScratchBuffer { get; set; }
-        public Uint32 ScratchBufferOffset { get; set; } = 0;
+        public Uint64 ScratchBufferOffset { get; set; } = 0;
         public RESOURCE_STATE_TRANSITION_MODE ScratchBufferTransitionMode { get; set; } = RESOURCE_STATE_TRANSITION_MODE.RESOURCE_STATE_TRANSITION_MODE_NONE;
         public Bool Update { get; set; } = false;
